OTP-Verification-rest-api/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
